Rewrites the package to use a macro-based approach.
The new approach should be more intuitive to write insertion code with.
Users use `splash-buffer-define` to create functions/variables needed by
a splash buffer instead of using the maste `splash-buffer` function to
create and/or switch to them.

A number of configuration parameters have been excised in the process,
and a default mode has been added for the likely use case of using the
`widgets` package to create buttons, etc.

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>